+++ features/messages/impl/src/main/kotlin/io/element/android/features/messages/impl/threads/list/ThreadsListPresenter.kt
@@ -20,13 +20,10 @@ import io.element.android.features.messages.impl.utils.messagesummary.MessageSum
 import io.element.android.libraries.architecture.Presenter
 import io.element.android.libraries.dateformatter.api.DateFormatter
 import io.element.android.libraries.dateformatter.api.DateFormatterMode
-import io.element.android.libraries.designsystem.components.avatar.AvatarData
 import io.element.android.libraries.designsystem.components.avatar.AvatarSize
-import io.element.android.libraries.matrix.api.core.RoomId
 import io.element.android.libraries.matrix.api.room.JoinedRoom
 import io.element.android.libraries.matrix.api.room.threads.ThreadListPaginationStatus
 import io.element.android.libraries.matrix.ui.model.getAvatarData
-import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.flow.onStart
@@ -128,18 +125,4 @@ class ThreadsListPresenter(
             eventSink = ::handleEvent,
         )
     }
-}
-
-data class ThreadsListState(
-    val roomId: RoomId,
-    val roomName: String,
-    val roomAvatarUrl: String?,
-    val isRoomTombstoned: Boolean,
-    val heroes: ImmutableList<AvatarData>,
-    val threads: ImmutableList<ThreadListRowItem>,
-    val eventSink: (ThreadsListEvent) -> Unit,
-)
-
-sealed interface ThreadsListEvent {
-    data object Paginate : ThreadsListEvent
 }
